requirements.txt
python-telegram-bot==20.7
openai==1.3.5
reportlab==4.0.7
PyPDF2==3.0.1
python-docx==0.8.11
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
gunicorn==21.2.0
uvicorn==0.24.0
fastapi==0.104.1
pydantic==2.5.0
asyncpg==0.29.0